Fix compilation warnings with -Wconversion

+++ esp32/main/tap.c
@@ -49,7 +49,7 @@ void tap__chunk(struct tap *t, const uint8_t *data, int len)
     //ESP_LOGI(TAG, "Parse chunk len %d state %d", len, t->state);
 
     if (t->initial_delay) {
-        vTaskDelay(t->initial_delay/portTICK_RATE_MS);
+        vTaskDelay((int)t->initial_delay/portTICK_RATE_MS);
         t->initial_delay = 0;
     }
